Rename ModelizerModelsView to ModelsView
